move replica to replica.h
